openapi endpoint changed to be a script
